// src/KRINT.Frontend/src/app/api/model/backupScheduleDto.ts
/**
 * KRINT.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BackupScheduleDto { 
    id: string;
    instanceId: string;
    cronExpression: string;
    description: string;
    enabled: boolean;
    lastRunAt?: string | null;
    lastStatus?: string | null;
    lastError?: string | null;
    nextRunAt?: string | null;
    createdAt: string;
}